Create Confusion Through Substitution
To use the SBox to transform the input, the number is converted to binary then to hexadecimal and mapped with the SBox. To get the plaintext we use the inverse SBox and map the values to get original matrix. Convert the int values to caharcters and append in a string to get the flag.
Concept:
https://books.google.co.in/books?id=J1zSBQAAQBAJ&lpg=PA394&ots=cJD-kLRHOW&dq=convert%20bytes%20in%20a%20box%20to%20matrix&pg=PA395#v=onepage&q&f=false